requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.36
pydantic[email]==2.9.2
bcrypt==4.0.1
requests==2.32.5
python-multipart==0.0.12
psycopg2-binary==2.9.11
pytest==8.4.2
python-jose[cryptography]==3.5.0
passlib==1.7.4